Fix: Crashes for Unlinked Text Links
In the UI, it is possible for the new text link field to yield what looks like a link structure without any linking information, i.e. a media or web link without a url property, or a document link without an id/uid etc.

This patch updates the factory to test for these 'broken' links and either return empty fragments or string fragments